refactor(engine): back file.suggest with a TtlCache-based FileSuggestService
